requirements.txt
fastmcp>=0.1.0
uvicorn>=0.30.0
pyyaml>=6.0